+++ src/banner.cpp
@@ -27,7 +27,7 @@ void print_lindblad_banner_once() {
 
         std::cout
             << "\n"
-            << "    Welcome to Lindblad Quantum Toolkit R.1.8.1 | CLI startup\n"
+            << "    Welcome to Lindblad Quantum Toolkit R.1.8.2 | CLI startup\n"
             << "__________________________________________________________________________________________________________________________________________________\n\n"
             << "              _     <─. (`─')_  _(`─')   <─.(`─')           (`─')  _ _(`─')             \n"
             << "      <─.    (_)       ╲( OO) )( (OO ).─> __( OO)    <─.    (OO ).─╱( (OO ).─>          \n"
@@ -58,7 +58,7 @@ void print_lindblad_exit_banner() {
     std::cout
         << "\n"
 
-        << "    Thank you for using Lindblad Quantum Toolkit R.1.8.1\n"
+        << "    Thank you for using Lindblad Quantum Toolkit R.1.8.2\n"
         << "__________________________________________________________________________________________________________________________________________________\n\n"
         << "              _     <─. (`─')_  _(`─')   <─.(`─')           (`─')  _ _(`─')             \n"
         << "      <─.    (_)       ╲( OO) )( (OO ).─> __( OO)    <─.    (OO ).─╱( (OO ).─>          \n"